REF: Use ArmaProcess.generate_sample instead of ArmaFft
Move armafft-specific test to an armafft-specific test file.  After this, non-sandbox does not import armafft
